Slim down dependencies in settings page
Jobs are disabled for now, so this also disables the buttons for
triggering those jobs.
